+++ backend/src/models/market_data.rs
@@ -1,0 +1,1 @@
+pub mod providers;
