+++ backend/src/models/market_data/providers/jquants/response.rs
